+++ backend/src/actors/Host.cpp
@@ -129,7 +129,7 @@ void Host::handle_region_request(const websocketpp::connection_hdl hdl,
         region.hOffset = 0;
         region.vOffset = 0;
 
-        region.maxIteration = request["region"]["maxIteration"].GetUint();
+        region.maxIteration = (unsigned short int) request["region"]["maxIteration"].GetUint();
         region.validation = request["region"]["validation"].GetInt();
         region.guaranteedDivisor = request["region"]["guaranteedDivisor"].GetUint();
     } catch (std::out_of_range &e) {
@@ -397,8 +397,8 @@ void Host::init(int world_rank, int world_size) {
             std::cout << "Host: MPI communication with Worker " << workerInfo.rank << " took approximately " << mpiCommunicationTime[workerInfo.rank - 1] << " microseconds." << std::endl;
         
             // Extract "worker_data" from the received message
-            int* worker_data = new int[region_size];
-            std::memcpy(worker_data, recv + sizeof(WorkerInfo), region_size * sizeof(int));
+            unsigned short int* worker_data = new unsigned short int[region_size];
+            std::memcpy(worker_data, recv + sizeof(WorkerInfo), region_size * sizeof(unsigned short int));
             std::cout << "Host: Receive from Worker " << workerInfo.rank << " complete." << std::endl;
         
             // Fill "region_data"
